Adapter to new lora test structure
